Inclusao do mongoDB no projeto

+++ qslog-back/src/qsLog/1 - Domains/Users/Repository/IUserRepository.cs
@@ -7,7 +7,7 @@ namespace qsLog.Domains.Users.Repository
 {
     public interface IUserRepository: IRepository<User, Guid>
     {
-        int Count();
+        long Count();
         IEnumerable<User> ListAll();
         Task<User> GetByUserName(string userName);
         bool ExistsUserName(string userName);
